incorporate minor changes from comments
