--- dashboard/src/components/settings/signInMethods/ConfirmProvidersUpdatedForm/index.ts
@@ -1,2 +1,0 @@
-export * from './ConfirmProvidersUpdatedForm';
-export { default } from './ConfirmProvidersUpdatedForm';
